// backend/src/types/benefit.ts
export type Benefit = {
    _id: number;
    short_title: string;
    long_title: string;
    short_description: string;
    long_description: string;
    thumbnail: string;
    contract_address: string;
    token_id: string;
    valid_from: string;
    valid_to: string;
    url: string;
    action_date: string;
};